ENGR00234042 Fix file over 2G size can not be written
The off_t in bionic is defined as int which is 32bits.
But in ntfs-3g, the off_t is required to define as
off64_t or loff_t 64bits.
Before this change, the fuse file write operation
with 32bit offset will failed when it's exceed 2G:
0x80000000. As signed integer it will be take as a
negtive value, after convert it to s64 in ntfs_fuse_write(),
the s64 offset param value will be 0xFFFFFFFFFF80000000.
Then causes offset over boundary and write failed.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>